+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the "note-organizer" extension will be documented in this file.
 
-## [Unreleased]
+## [0.2.1]
  - Project and note scanning: find all of the project instead of only those with notes.
  - Toggle to hide and show empty project in tree data.